Added separate files for extensions for better readability
