+++ PICommandLineTool/Extensions/NSRegularExpression+Extension.swift
@@ -1,0 +1,38 @@
+//
+//  NSRegularExpression+Extension.swift
+//  PICommandLineTool
+//
+//  Created by Pavan on 11/18/19.
+//  Copyright © 2019 Pavan Itagi. All rights reserved.
+//
+
+import Foundation
+
+extension NSRegularExpression {
+    convenience init(_ pattern: String) {
+        do {
+            try self.init(pattern: pattern)
+        } catch {
+            preconditionFailure("Illegal regular expression: \(pattern).")
+        }
+    }
+    
+    convenience init(_ pattern: String, options: NSRegularExpression.Options = NSRegularExpression.Options.caseInsensitive) {
+        do {
+            try self.init(pattern: pattern, options: options)
+        } catch {
+            preconditionFailure("Illegal regular expression: \(pattern).")
+        }
+    }
+}
+
+extension NSRegularExpression {
+    func matches(_ string: String) -> Bool {
+        let range = NSRange(location: 0, length: string.utf16.count)
+        return firstMatch(in: string, options: [], range: range) != nil
+    }
+    
+    func matches(in string: String, options: NSRegularExpression.MatchingOptions = []) -> [NSTextCheckingResult] {
+        return matches(in: string, options: options, range: NSRange(location: 0, length: string.utf16.count))
+    }
+}
