+++ PICommandLineTool/Extensions/String+extension.swift
@@ -1,0 +1,37 @@
+//
+//  String+extension.swift
+//  PICommandLineTool
+//
+//  Created by Pavan on 11/18/19.
+//  Copyright © 2019 Pavan Itagi. All rights reserved.
+//
+
+import Foundation
+
+extension String {
+    var verifyUrl: Bool {
+        get {
+            let url = URL(string: self)
+
+            if url == nil || NSData(contentsOf: url!) == nil {
+                return false
+            } else {
+                return true
+            }
+        }
+    }
+    
+    var validFileURL: Bool {
+        get {
+            let url = URL(fileURLWithPath: self)
+            return FileManager.default.fileExists(atPath: url.absoluteString)
+        }
+    }
+    
+    // Example : "hat" ~= "[a-z]at"
+    static func ~= (lhs: String, rhs: String) -> Bool {
+        guard let regex = try? NSRegularExpression(pattern: rhs) else { return false }
+        let range = NSRange(location: 0, length: lhs.utf16.count)
+        return regex.firstMatch(in: lhs, options: [], range: range) != nil
+    }
+}
